Move models [fields] -> [model]
