+++ client/src/commonMain/kotlin/client/viewModel/RateMovieViewModel.kt
@@ -5,7 +5,7 @@ import domain.FindMovie
 import domain.stats.AddMovieToWatchlist
 import domain.stats.RateMovie
 import entities.TmdbId
-import entities.field.UserRating
+import entities.model.UserRating
 import entities.movies.Movie
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
