+++ database/src/commonMain/kotlin/database/di.kt
@@ -17,7 +17,7 @@ import database.stats.Watchlist
 import entities.IntId
 import entities.TmdbId
 import entities.TmdbStringId
-import entities.field.Name
+import entities.model.Name
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
@@ -120,9 +120,9 @@ val databaseModule = module {
     factory<ColumnAdapter<TmdbId, Long>>(tmdbIdAdapter) { TmdbIdAdapter() }
     factory<ColumnAdapter<TmdbStringId, String>>(tmdbStringIdAdapter) { TmdbStringIdAdapter() }
     factory<ColumnAdapter<Name, String>>(nameAdapter) { NameAdapter() }
-    factory<ColumnAdapter<entities.field.Video.Site, String>>(siteAdapter) { SiteAdapter() }
+    factory<ColumnAdapter<entities.model.Video.Site, String>>(siteAdapter) { SiteAdapter() }
     factory<ColumnAdapter<StatType, String>>(statTypeAdapter) { StatTypeAdapter() }
-    factory<ColumnAdapter<entities.field.Video.Type, String>>(videoTypeAdapter) { VideoTypeAdapter() }
+    factory<ColumnAdapter<entities.model.Video.Type, String>>(videoTypeAdapter) { VideoTypeAdapter() }
 }
 
 internal const val DATABASE_FILE_NAME = "cinescout.db"
