+++ database/src/commonMain/sqldelight/database/movies/genre.sq
@@ -1,6 +1,6 @@
 import entities.IntId;
 import entities.TmdbId;
-import entities.field.Name;
+import entities.model.Name;
 
 CREATE TABLE IF NOT EXISTS genre (
   id INTEGER as IntId NOT NULL PRIMARY KEY AUTOINCREMENT,
