+++ movies/remote/tmdb/src/commonMain/kotlin/movies/remote/tmdb/TmdbRemoteMovieSourceImpl.kt
@@ -4,12 +4,12 @@ import com.soywiz.klock.DateFormat
 import com.soywiz.klock.parse
 import entities.TmdbId
 import entities.TmdbStringId
-import entities.field.Actor
-import entities.field.CommunityRating
-import entities.field.Genre
-import entities.field.Name
-import entities.field.TmdbImageUrl
-import entities.field.Video
+import entities.model.Actor
+import entities.model.CommunityRating
+import entities.model.Genre
+import entities.model.Name
+import entities.model.TmdbImageUrl
+import entities.model.Video
 import entities.movies.DiscoverParams
 import entities.movies.Movie
 import io.ktor.client.features.ClientRequestException
