+++ stats/local/src/commonMain/kotlin/stats/local/LocalStatSourceImpl.kt
@@ -14,13 +14,13 @@ import database.stats.StatType
 import database.stats.WatchlistQueries
 import entities.IntId
 import entities.TmdbId
-import entities.field.Actor
-import entities.field.CommunityRating
-import entities.field.FiveYearRange
-import entities.field.Genre
-import entities.field.TmdbImageUrl
-import entities.field.UserRating
-import entities.field.Video
+import entities.model.Actor
+import entities.model.CommunityRating
+import entities.model.FiveYearRange
+import entities.model.Genre
+import entities.model.TmdbImageUrl
+import entities.model.UserRating
+import entities.model.Video
 import entities.movies.Movie
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
