+++ stats/local/src/commonTest/kotlin/stats/local/double/MockGenreQueries.kt
@@ -3,7 +3,7 @@ package stats.local.double
 import database.movies.GenreQueries
 import entities.IntId
 import entities.TmdbId
-import entities.field.Name
+import entities.model.Name
 import io.mockk.every
 import io.mockk.mockk
