+++ stats/local/src/commonTest/kotlin/stats/local/double/MockStatQueries.kt
@@ -5,7 +5,7 @@ import database.stats.StatQueries
 import database.stats.StatType
 import entities.IntId
 import entities.TmdbId
-import entities.field.Name
+import entities.model.Name
 import io.mockk.every
 import io.mockk.mockk
